[#842] Phase 1: ClaudePhaseExecutor 추출 + PhaseExecutor 인터페이스 정의
